fix the subgraph from a graft

+++ src/mappings/position-manager.ts
@@ -8,7 +8,7 @@ import {
 } from '../types/NonfungiblePositionManager/NonfungiblePositionManager'
 import { Position, PositionSnapshot, Token } from '../types/schema'
 import { ADDRESS_ZERO, factoryContract, ZERO_BD, ZERO_BI } from '../utils/constants'
-import { Address, BigInt, ethereum } from '@graphprotocol/graph-ts'
+import { Address, BigInt, BigDecimal, ethereum } from '@graphprotocol/graph-ts'
 import { convertTokenToDecimal, loadTransaction } from '../utils'
 
 function getPosition(event: ethereum.Event, tokenId: BigInt): Position | null {
@@ -73,8 +73,8 @@ function savePositionSnapshot(position: Position, event: ethereum.Event): void {
   positionSnapshot.depositedToken1 = position.depositedToken1
   positionSnapshot.withdrawnToken0 = position.withdrawnToken0
   positionSnapshot.withdrawnToken1 = position.withdrawnToken1
-  positionSnapshot.collectedFeesToken0 = position.collectedFeesToken0
-  positionSnapshot.collectedFeesToken1 = position.collectedFeesToken1
+  positionSnapshot.collectedFeesToken0 = position.collectedFeesToken0 as BigDecimal
+  positionSnapshot.collectedFeesToken1 = position.collectedFeesToken1 as BigDecimal
   positionSnapshot.transaction = loadTransaction(event).id
   positionSnapshot.feeGrowthInside0LastX128 = position.feeGrowthInside0LastX128
   positionSnapshot.feeGrowthInside1LastX128 = position.feeGrowthInside1LastX128
